Refactor examples to avoid code duplication.
